+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.6.2] - 2026-06-11
+
+### Added
+
+- Mask cleanup using morphological opening and closing (unconfigurable for now)
+- Option to limit the interpolation applied
+
+### Changed
+
+- Hue tolerance for marker finding (15 to 10)
+
+### Fixed
+
+- Wrong changelog in the last releases
+- Various tracking and postprocessing bugs
+
 ## [v0.6.1] - 2026-06-10
 
 ### Added